Refactor: create options class
